Create a utils file for animations
